Fixed the compilation errors due to recent backend changes.
