+++ backend/src/apiserver/storage/BUILD.bazel
@@ -22,15 +22,12 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//backend/api:go_default_library",
-        "//backend/src/apiserver/client:go_default_library",
         "//backend/src/apiserver/common:go_default_library",
         "//backend/src/apiserver/list:go_default_library",
         "//backend/src/apiserver/metadata:go_default_library",
         "//backend/src/apiserver/model:go_default_library",
         "//backend/src/common/util:go_default_library",
         "@com_github_argoproj_argo//pkg/apis/workflow/v1alpha1:go_default_library",
-        "@com_github_argoproj_argo//pkg/client/clientset/versioned/typed/workflow/v1alpha1:go_default_library",
-        "@com_github_cenkalti_backoff//:go_default_library",
         "@com_github_ghodss_yaml//:go_default_library",
         "@com_github_go_sql_driver_mysql//:go_default_library",
         "@com_github_golang_glog//:go_default_library",
@@ -40,13 +37,10 @@ go_library(
         "@com_github_minio_minio_go//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_vividcortex_mysqlerr//:go_default_library",
-<<<<<<< HEAD
         "@google_ml_metadata//ml_metadata/metadata_store:metadata_store_go",  # keep
         "@google_ml_metadata//ml_metadata/proto:metadata_store_go_proto",  # keep
         "@google_ml_metadata//ml_metadata/proto:metadata_store_service_go_proto",  # keep
         "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
-=======
->>>>>>> e0c63ff... Using the backoff package for retries.
         "@io_k8s_apimachinery//pkg/types:go_default_library",
         "@io_k8s_apimachinery//pkg/util/json:go_default_library",
     ],
